[JPA 1.1] Fix CT to use complete persistence units for tests that require them
The previous tests were configured with persistence units that named a driver, but no URL. As a result the persistence units were incomplete and the tests that expected an EMF were wrong. These persistence units now use an in-memory H2 URL.

Note that a bug in the Oracle Gemini bundle (used as the RI) prevents any real DataSourceFactory from being used to create the Database Driver. As a result this commit also upgrades the RI to the latest release of Eclipse Gemini JPA (version 1.1.0). This version also has a bug which means that JPATestCase#testDataSourceFactoryUnregistration() fails.